Remove Algolia CSS and update documentation structure
The Algolia CSS file was removed, impacting the configuration file as well. Moreover, the documentation's structure was heavily updated, removing unnecessary sections and refining existing ones for clarity. Sections related to community and contributions are emphasized now. The standardization of the table of contents and captions was also a prominent part of the update.