Add an entry point for unit testing
I didn’t find any unit tests, so I’m proposing that tests be written in a new `tests/` which I understand to be a common convention.

I’ve added a single, tiny unit test that demonstrates that the convenience import that’s supported for other modules isn’t there for `OllamaLocal`.

To run the unit tests, you just have to run this command:

    python -m unittest discover tests/

With this commit it will fail with this output to the console:

    E
    ======================================================================
    ERROR: test_convenience_import_of_dsp_ollama_client (test_dsp.DspModuleImportsTest.test_convenience_import_of_dsp_ollama_client)
    ----------------------------------------------------------------------
    Traceback (most recent call last):
      File "/Users/kevin/Developer/stanfordnlp/dspy/tests/test_dsp.py", line 6, in test_convenience_import_of_dsp_ollama_client
        from dsp import OllamaLocal
    ImportError: cannot import name 'OllamaLocal' from 'dsp' (/Users/kevin/Developer/stanfordnlp/dspy/dsp/__init__.py)

    ----------------------------------------------------------------------
    Ran 1 test in 0.479s

    FAILED (errors=1)