fix(player.py): Fix syntax error in player file which downloads into Python312\Lib\site-packages\dspy upon installation
This "player.py" file, though mentioned by name in the 0.0.3 dspy documentation, is not initially present in the github repo itself but downloads into Python312\Lib\site-packages\dspy when DSPy is installed. This fixes a recurring error I have encountered locally, in conda, and on Hugging Face Spaces.
I hope this is the correct way to submit this fix, despite the location of the player.py file.
Error states on Hugging Face as:
'''
Traceback (most recent call last):
  File "/home/user/app/app.py", line 24, in <module>
    from dspy.dspy.agents import Agent
  File "/home/user/.local/lib/python3.10/site-packages/dspy/__init__.py", line 12, in <module>
    from dspy.player import Player
  File "/home/user/.local/lib/python3.10/site-packages/dspy/player.py", line 38
    print 'Too many generators to append another'
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
SyntaxError: Missing parentheses in call to 'print'. Did you mean print(...)?
'''