Enhance README "Some other examples"
Modified so that the "benchmark on a bunch of different tasks" sends the reader to a README.md file explaining the benchmarking process, instead of just dropping them into the directory on GitHub.

Added a link to ArXiV for "The Unreasonable Effectiveness..." article, so that now the interested reader can jump directly to ArXiV instead of having to go through the Register article. The link to the Register article *is* retained, though.