Make explicit how to get LLM response in first examples
The first examples should be self-contained and copy-pastable for easy use for first time users. But not all of the first examples actually capture the response from the LLM or print them out. Here, I've updated those examples so that a first time user can simply copy and paste the example script and see the output produced.